refactored rabbit container init to its own fixture
